[refactor] Extract TableWriteOperator class
