+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/sink/RowDataStoreWriteOperator.java
@@ -20,14 +20,12 @@ package org.apache.paimon.flink.sink;
 
 import org.apache.paimon.flink.FlinkRowWrapper;
 import org.apache.paimon.flink.log.LogWriteCallback;
-import org.apache.paimon.options.Options;
 import org.apache.paimon.table.FileStoreTable;
 import org.apache.paimon.table.sink.SinkRecord;
 
 import org.apache.flink.api.common.functions.util.FunctionUtils;
 import org.apache.flink.api.common.state.CheckpointListener;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.runtime.state.StateInitializationContext;
 import org.apache.flink.runtime.state.StateSnapshotContext;
 import org.apache.flink.streaming.api.functions.sink.SinkFunction;
 import org.apache.flink.streaming.api.graph.StreamConfig;
@@ -45,20 +43,12 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.List;
 
-import static org.apache.paimon.flink.FlinkConnectorOptions.SINK_USE_MANAGED_MEMORY;
-
 /** A {@link PrepareCommitOperator} to write {@link RowData}. Record schema is fixed. */
-public class RowDataStoreWriteOperator extends PrepareCommitOperator<RowData, Committable> {
+public class RowDataStoreWriteOperator extends TableWriteOperator<RowData> {
 
     private static final long serialVersionUID = 3L;
 
-    private final FileStoreTable table;
     @Nullable private final LogSinkFunction logSinkFunction;
-    private final StoreSinkWrite.Provider storeSinkWriteProvider;
-    private final String initialCommitUser;
-
-    private transient StoreSinkWriteState state;
-    private transient StoreSinkWrite write;
     private transient SimpleContext sinkContext;
     @Nullable private transient LogWriteCallback logCallback;
 
@@ -70,11 +60,8 @@ public class RowDataStoreWriteOperator extends PrepareCommitOperator<RowData, Co
             @Nullable LogSinkFunction logSinkFunction,
             StoreSinkWrite.Provider storeSinkWriteProvider,
             String initialCommitUser) {
-        super(Options.fromMap(table.options()));
-        this.table = table;
+        super(table, storeSinkWriteProvider, initialCommitUser);
         this.logSinkFunction = logSinkFunction;
-        this.storeSinkWriteProvider = storeSinkWriteProvider;
-        this.initialCommitUser = initialCommitUser;
     }
 
     @Override
@@ -89,42 +76,8 @@ public class RowDataStoreWriteOperator extends PrepareCommitOperator<RowData, Co
     }
 
     @Override
-    public void initializeState(StateInitializationContext context) throws Exception {
-        super.initializeState(context);
-
-        // Each job can only have one user name and this name must be consistent across restarts.
-        // We cannot use job id as commit user name here because user may change job id by creating
-        // a savepoint, stop the job and then resume from savepoint.
-        String commitUser =
-                StateUtils.getSingleValueFromState(
-                        context, "commit_user_state", String.class, initialCommitUser);
-
-        RowDataChannelComputer channelComputer =
-                new RowDataChannelComputer(table.schema(), logSinkFunction != null);
-        channelComputer.setup(getRuntimeContext().getNumberOfParallelSubtasks());
-        state =
-                new StoreSinkWriteState(
-                        context,
-                        (tableName, partition, bucket) ->
-                                channelComputer.channel(partition, bucket)
-                                        == getRuntimeContext().getIndexOfThisSubtask());
-
-        Options options = Options.fromMap(table.options());
-        if (options.get(SINK_USE_MANAGED_MEMORY) && memoryPool == null) {
-            throw new RuntimeException("Memory pool must be initialized first.");
-        }
-
-        write =
-                storeSinkWriteProvider.provide(
-                        table,
-                        commitUser,
-                        state,
-                        getContainingTask().getEnvironment().getIOManager(),
-                        memoryPool);
-
-        if (logSinkFunction != null) {
-            StreamingFunctionUtils.restoreFunctionState(context, logSinkFunction);
-        }
+    protected boolean containLogSystem() {
+        return logSinkFunction != null;
     }
 
     @Override
@@ -172,9 +125,6 @@ public class RowDataStoreWriteOperator extends PrepareCommitOperator<RowData, Co
     public void snapshotState(StateSnapshotContext context) throws Exception {
         super.snapshotState(context);
 
-        write.snapshotState();
-        state.snapshotState();
-
         if (logSinkFunction != null) {
             StreamingFunctionUtils.snapshotFunctionState(
                     context, getOperatorStateBackend(), logSinkFunction);
@@ -194,7 +144,6 @@ public class RowDataStoreWriteOperator extends PrepareCommitOperator<RowData, Co
     public void close() throws Exception {
         super.close();
 
-        write.close();
         if (logSinkFunction != null) {
             FunctionUtils.closeFunction(logSinkFunction);
         }
@@ -221,7 +170,7 @@ public class RowDataStoreWriteOperator extends PrepareCommitOperator<RowData, Co
     @Override
     protected List<Committable> prepareCommit(boolean doCompaction, long checkpointId)
             throws IOException {
-        List<Committable> committables = write.prepareCommit(doCompaction, checkpointId);
+        List<Committable> committables = super.prepareCommit(doCompaction, checkpointId);
 
         if (logCallback != null) {
             try {
